Powerups have been added.
The "Bouncy Ball" power is still too fast for it to be playable. The biggest bug right now is that the board is not resetting properly to its original state after a power is called. This is likely because of issues with the function reset() which is supposed to reset the board after any one of the three powerups has been used for 10 seconds.